# supabase/schema-openapi.json
{
  "swagger": "2.0",
  "info": {
    "description": "",
    "title": "standard public schema",
    "version": "14.1"
  },
  "host": "wjcbobcqyqdkludsbqgf.supabase.co:443",
  "basePath": "/",
  "schemes": [
    "https"
  ],
  "consumes": [
    "application/json",
    "application/vnd.pgrst.object+json;nulls=stripped",
    "application/vnd.pgrst.object+json",
    "text/csv"
  ],
  "produces": [
    "application/json",
    "application/vnd.pgrst.object+json;nulls=stripped",
    "application/vnd.pgrst.object+json",
    "text/csv"
  ],
  "paths": {
...
    "/forum": {
      "get": {
        "parameters": [
          {
            "$ref": "#/parameters/rowFilter.forum.id"
          },
          {
            "$ref": "#/parameters/rowFilter.forum.titulo"
          },
          {
            "$ref": "#/parameters/rowFilter.forum.descricao"
          },
          {
            "$ref": "#/parameters/rowFilter.forum.conteudo"
          },
          {
            "$ref": "#/parameters/rowFilter.forum.autorid"
          },
          {
            "$ref": "#/parameters/rowFilter.forum.disciplinaid"
          },
          {
            "$ref": "#/parameters/rowFilter.forum.turmaid"
          },
          {
            "$ref": "#/parameters/rowFilter.forum.datapostagem"
          },
          {
            "$ref": "#/parameters/rowFilter.forum.datacriacao"
          },
          {
            "$ref": "#/parameters/rowFilter.forum.dataatualizacao"
          },
          {
            "$ref": "#/parameters/select"
          },
          {
            "$ref": "#/parameters/order"
          },
          {
            "$ref": "#/parameters/range"
          },
          {
            "$ref": "#/parameters/rangeUnit"
          },
          {
            "$ref": "#/parameters/offset"
          },
          {
            "$ref": "#/parameters/limit"
          },
          {
            "$ref": "#/parameters/preferCount"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "items": {
                "$ref": "#/definitions/forum"
              },
              "type": "array"
            }
          },
          "206": {
            "description": "Partial Content"
          }
        },
        "tags": [
          "forum"
        ]
      },
      "post": {
        "parameters": [
          {
            "$ref": "#/parameters/body.forum"
          },
          {
            "$ref": "#/parameters/select"
          },
          {
            "$ref": "#/parameters/preferPost"
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          }
        },
        "tags": [
          "forum"
        ]
      },
      "delete": {
        "parameters": [
          {
            "$ref": "#/parameters/rowFilter.forum.id"
          },
          {
            "$ref": "#/parameters/rowFilter.forum.titulo"
          },
          {
            "$ref": "#/parameters/rowFilter.forum.descricao"
          },
          {
            "$ref": "#/parameters/rowFilter.forum.conteudo"
          },
          {
            "$ref": "#/parameters/rowFilter.forum.autorid"
          },
          {
            "$ref": "#/parameters/rowFilter.forum.disciplinaid"
          },
          {
            "$ref": "#/parameters/rowFilter.forum.turmaid"
          },
          {
            "$ref": "#/parameters/rowFilter.forum.datapostagem"
          },
          {
            "$ref": "#/parameters/rowFilter.forum.datacriacao"
          },
          {
            "$ref": "#/parameters/rowFilter.forum.dataatualizacao"
          },
          {
            "$ref": "#/parameters/preferReturn"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "tags": [
          "forum"
        ]
      },
      "patch": {
        "parameters": [
          {
            "$ref": "#/parameters/rowFilter.forum.id"
          },
          {
            "$ref": "#/parameters/rowFilter.forum.titulo"
          },
          {
            "$ref": "#/parameters/rowFilter.forum.descricao"
          },
          {
            "$ref": "#/parameters/rowFilter.forum.conteudo"
          },
          {
            "$ref": "#/parameters/rowFilter.forum.autorid"
          },
          {
            "$ref": "#/parameters/rowFilter.forum.disciplinaid"
          },
          {
            "$ref": "#/parameters/rowFilter.forum.turmaid"
          },
          {
            "$ref": "#/parameters/rowFilter.forum.datapostagem"
          },
          {
            "$ref": "#/parameters/rowFilter.forum.datacriacao"
          },
          {
            "$ref": "#/parameters/rowFilter.forum.dataatualizacao"
          },
          {
            "$ref": "#/parameters/body.forum"
          },
          {
            "$ref": "#/parameters/preferReturn"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "tags": [
          "forum"
        ]
      }
    },
...
    "/disciplinas": {
      "get": {
        "parameters": [
          {
            "$ref": "#/parameters/rowFilter.disciplinas.id"
          },
          {
            "$ref": "#/parameters/rowFilter.disciplinas.nome"
          },
          {
            "$ref": "#/parameters/rowFilter.disciplinas.codigo"
          },
          {
            "$ref": "#/parameters/rowFilter.disciplinas.cursoid"
          },
          {
            "$ref": "#/parameters/rowFilter.disciplinas.cargahoraria"
          },
          {
            "$ref": "#/parameters/rowFilter.disciplinas.ementa"
          },
          {
            "$ref": "#/parameters/rowFilter.disciplinas.objetivos"
          },
          {
            "$ref": "#/parameters/rowFilter.disciplinas.periodo"
          },
          {
            "$ref": "#/parameters/rowFilter.disciplinas.situacao"
          },
          {
            "$ref": "#/parameters/rowFilter.disciplinas.datacriacao"
          },
          {
            "$ref": "#/parameters/rowFilter.disciplinas.dataatualizacao"
          },
          {
            "$ref": "#/parameters/rowFilter.disciplinas.numero_id"
          },
          {
            "$ref": "#/parameters/select"
          },
          {
            "$ref": "#/parameters/order"
          },
          {
            "$ref": "#/parameters/range"
          },
          {
            "$ref": "#/parameters/rangeUnit"
          },
          {
            "$ref": "#/parameters/offset"
          },
          {
            "$ref": "#/parameters/limit"
          },
          {
            "$ref": "#/parameters/preferCount"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "items": {
                "$ref": "#/definitions/disciplinas"
              },
              "type": "array"
            }
          },
          "206": {
            "description": "Partial Content"
          }
        },
        "tags": [
          "disciplinas"
        ]
      },
      "post": {
        "parameters": [
          {
            "$ref": "#/parameters/body.disciplinas"
          },
          {
            "$ref": "#/parameters/select"
          },
          {
            "$ref": "#/parameters/preferPost"
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          }
        },
        "tags": [
          "disciplinas"
        ]
      },
      "delete": {
        "parameters": [
          {
            "$ref": "#/parameters/rowFilter.disciplinas.id"
          },
          {
            "$ref": "#/parameters/rowFilter.disciplinas.nome"
          },
          {
            "$ref": "#/parameters/rowFilter.disciplinas.codigo"
          },
          {
            "$ref": "#/parameters/rowFilter.disciplinas.cursoid"
          },
          {
            "$ref": "#/parameters/rowFilter.disciplinas.cargahoraria"
          },
          {
            "$ref": "#/parameters/rowFilter.disciplinas.ementa"
          },
          {
            "$ref": "#/parameters/rowFilter.disciplinas.objetivos"
          },
          {
            "$ref": "#/parameters/rowFilter.disciplinas.periodo"
          },
          {
            "$ref": "#/parameters/rowFilter.disciplinas.situacao"
          },
          {
            "$ref": "#/parameters/rowFilter.disciplinas.datacriacao"
          },
          {
            "$ref": "#/parameters/rowFilter.disciplinas.dataatualizacao"
          },
          {
            "$ref": "#/parameters/rowFilter.disciplinas.numero_id"
          },
          {
            "$ref": "#/parameters/preferReturn"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "tags": [
          "disciplinas"
        ]
      },
      "patch": {
        "parameters": [
          {
            "$ref": "#/parameters/rowFilter.disciplinas.id"
          },
          {
            "$ref": "#/parameters/rowFilter.disciplinas.nome"
          },
          {
            "$ref": "#/parameters/rowFilter.disciplinas.codigo"
          },
          {
            "$ref": "#/parameters/rowFilter.disciplinas.cursoid"
          },
          {
            "$ref": "#/parameters/rowFilter.disciplinas.cargahoraria"
          },
          {
            "$ref": "#/parameters/rowFilter.disciplinas.ementa"
          },
          {
            "$ref": "#/parameters/rowFilter.disciplinas.objetivos"
          },
          {
            "$ref": "#/parameters/rowFilter.disciplinas.periodo"
          },
          {
            "$ref": "#/parameters/rowFilter.disciplinas.situacao"
          },
          {
            "$ref": "#/parameters/rowFilter.disciplinas.datacriacao"
          },
          {
            "$ref": "#/parameters/rowFilter.disciplinas.dataatualizacao"
          },
          {
            "$ref": "#/parameters/rowFilter.disciplinas.numero_id"
          },
          {
            "$ref": "#/parameters/body.disciplinas"
          },
          {
            "$ref": "#/parameters/preferReturn"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "tags": [
          "disciplinas"
        ]
      }
    },
...
  },
  "definitions": {
...
    "forum": {
      "required": [
        "id",
        "titulo",
        "autorid"
      ],
      "properties": {
        "id": {
          "description": "Note:\nThis is a Primary Key.<pk/>",
          "format": "integer",
          "type": "integer"
        },
        "titulo": {
          "format": "character varying",
          "maxLength": 255,
          "type": "string"
        },
        "descricao": {
          "format": "text",
          "type": "string"
        },
        "conteudo": {
          "format": "text",
          "type": "string"
        },
        "autorid": {
          "description": "Note:\nThis is a Foreign Key to `usuarios.id`.<fk table='usuarios' column='id'/>",
          "format": "integer",
          "type": "integer"
        },
        "disciplinaid": {
          "description": "Note:\nThis is a Foreign Key to `disciplinas.id`.<fk table='disciplinas' column='id'/>",
          "format": "integer",
          "type": "integer"
        },
        "turmaid": {
          "description": "Note:\nThis is a Foreign Key to `turmas.id`.<fk table='turmas' column='id'/>",
          "format": "integer",
          "type": "integer"
        },
        "datapostagem": {
          "default": "CURRENT_TIMESTAMP",
          "format": "timestamp without time zone",
          "type": "string"
        },
        "datacriacao": {
          "default": "CURRENT_TIMESTAMP",
          "format": "timestamp without time zone",
          "type": "string"
        },
        "dataatualizacao": {
          "default": "CURRENT_TIMESTAMP",
          "format": "timestamp without time zone",
          "type": "string"
        }
      },
      "type": "object"
    },
...
    "disciplinas": {
      "required": [
        "id",
        "nome",
        "cursoid",
        "numero_id"
      ],
      "properties": {
        "id": {
          "description": "Note:\nThis is a Primary Key.<pk/>",
          "format": "integer",
          "type": "integer"
        },
        "nome": {
          "format": "character varying",
          "maxLength": 255,
          "type": "string"
        },
        "codigo": {
          "format": "character varying",
          "maxLength": 50,
          "type": "string"
        },
        "cursoid": {
          "description": "Note:\nThis is a Foreign Key to `cursos.id`.<fk table='cursos' column='id'/>",
          "format": "integer",
          "type": "integer"
        },
        "cargahoraria": {
          "format": "integer",
          "type": "integer"
        },
        "ementa": {
          "format": "text",
          "type": "string"
        },
        "objetivos": {
          "format": "text",
          "type": "string"
        },
        "periodo": {
          "format": "integer",
          "type": "integer"
        },
        "situacao": {
          "default": "ATIVO",
          "format": "character varying",
          "maxLength": 50,
          "type": "string"
        },
        "datacriacao": {
          "default": "CURRENT_TIMESTAMP",
          "format": "timestamp without time zone",
          "type": "string"
        },
        "dataatualizacao": {
          "default": "CURRENT_TIMESTAMP",
          "format": "timestamp without time zone",
          "type": "string"
        },
        "numero_id": {
          "format": "integer",
          "type": "integer"
        }
      },
      "type": "object"
    },
...
  },
  "parameters": {
...
    "preferReturn": {
      "name": "Prefer",
      "description": "Preference",
      "required": false,
      "enum": [
        "return=representation",
        "return=minimal",
        "return=none"
      ],
      "in": "header",
      "type": "string"
    },
    "preferCount": {
      "name": "Prefer",
      "description": "Preference",
      "required": false,
      "enum": [
        "count=none"
      ],
      "in": "header",
      "type": "string"
    },
    "preferPost": {
      "name": "Prefer",
      "description": "Preference",
      "required": false,
      "enum": [
        "return=representation",
        "return=minimal",
        "return=none",
        "resolution=ignore-duplicates",
        "resolution=merge-duplicates"
      ],
      "in": "header",
      "type": "string"
    },
    "select": {
      "name": "select",
      "description": "Filtering Columns",
      "required": false,
      "in": "query",
      "type": "string"
    },
...
    "order": {
      "name": "order",
      "description": "Ordering",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "range": {
      "name": "Range",
      "description": "Limiting and Pagination",
      "required": false,
      "in": "header",
      "type": "string"
    },
    "rangeUnit": {
      "name": "Range-Unit",
      "description": "Limiting and Pagination",
      "required": false,
      "default": "items",
      "in": "header",
      "type": "string"
    },
    "offset": {
      "name": "offset",
      "description": "Limiting and Pagination",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "limit": {
      "name": "limit",
      "description": "Limiting and Pagination",
      "required": false,
      "in": "query",
      "type": "string"
    },
...
    "body.forum": {
      "name": "forum",
      "description": "forum",
      "required": false,
      "in": "body",
      "schema": {
        "$ref": "#/definitions/forum"
      }
    },
    "rowFilter.forum.id": {
      "name": "id",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.forum.titulo": {
      "name": "titulo",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.forum.descricao": {
      "name": "descricao",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.forum.conteudo": {
      "name": "conteudo",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.forum.autorid": {
      "name": "autorid",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.forum.disciplinaid": {
      "name": "disciplinaid",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.forum.turmaid": {
      "name": "turmaid",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.forum.datapostagem": {
      "name": "datapostagem",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.forum.datacriacao": {
      "name": "datacriacao",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.forum.dataatualizacao": {
      "name": "dataatualizacao",
      "required": false,
      "in": "query",
      "type": "string"
    },
...
    "body.disciplinas": {
      "name": "disciplinas",
      "description": "disciplinas",
      "required": false,
      "in": "body",
      "schema": {
        "$ref": "#/definitions/disciplinas"
      }
    },
    "rowFilter.disciplinas.id": {
      "name": "id",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.disciplinas.nome": {
      "name": "nome",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.disciplinas.codigo": {
      "name": "codigo",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.disciplinas.cursoid": {
      "name": "cursoid",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.disciplinas.cargahoraria": {
      "name": "cargahoraria",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.disciplinas.ementa": {
      "name": "ementa",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.disciplinas.objetivos": {
      "name": "objetivos",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.disciplinas.periodo": {
      "name": "periodo",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.disciplinas.situacao": {
      "name": "situacao",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.disciplinas.datacriacao": {
      "name": "datacriacao",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.disciplinas.dataatualizacao": {
      "name": "dataatualizacao",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.disciplinas.numero_id": {
      "name": "numero_id",
      "required": false,
      "in": "query",
      "type": "string"
    },
...
  },
...
}